added language load on app component
